+++ shared/ai/src/user.rs
@@ -1,5 +1,6 @@
 pub struct Ship {}
 
+#[allow(clippy::new_without_default)]
 impl Ship {
     pub fn new() -> Ship {
         Ship {}